# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PoE2 Hideout Butler API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/refresh": {
      "post": {
        "tags": [
          "refresh"
        ],
        "summary": "Refresh snapshot data",
        "operationId": "refresh_api_refresh_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stashes": {
      "get": {
        "tags": [
          "stashes"
        ],
        "summary": "List stash tabs for a league",
        "operationId": "list_stashes_api_stashes_get",
        "parameters": [
          {
            "name": "league",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "League"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stashes/{tab_id}": {
      "get": {
        "tags": [
          "stashes"
        ],
        "summary": "Stash tab contents",
        "operationId": "get_stash_tab_api_stashes__tab_id__get",
        "parameters": [
          {
            "name": "tab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tab Id"
            }
          },
          {
            "name": "league",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "League"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashTab"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stashes/refresh": {
      "post": {
        "tags": [
          "stashes"
        ],
        "summary": "Refresh stash snapshot for a league",
        "operationId": "refresh_stash_tabs_api_stashes_refresh_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshStashesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Refresh Stash Tabs Api Stashes Refresh Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Item": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "inventory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory Id"
          },
          "w": {
            "type": "integer",
            "title": "W",
            "default": 1
          },
          "h": {
            "type": "integer",
            "title": "H",
            "default": 1
          },
          "x": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "type_line": {
            "type": "string",
            "title": "Type Line",
            "default": ""
          },
          "base_type": {
            "type": "string",
            "title": "Base Type",
            "default": ""
          },
          "rarity": {
            "type": "string",
            "title": "Rarity",
            "default": "Normal"
          },
          "ilvl": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ilvl"
          },
          "identified": {
            "type": "boolean",
            "title": "Identified",
            "default": true
          },
          "corrupted": {
            "type": "boolean",
            "title": "Corrupted",
            "default": false
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/ItemProperty"
            },
            "type": "array",
            "title": "Properties"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/ItemProperty"
            },
            "type": "array",
            "title": "Requirements"
          },
          "implicit_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Implicit Mods"
          },
          "explicit_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explicit Mods"
          },
          "rune_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rune Mods"
          },
          "enchant_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enchant Mods"
          },
          "crafted_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Crafted Mods"
          },
          "sockets": {
            "items": {
              "$ref": "#/components/schemas/Socket"
            },
            "type": "array",
            "title": "Sockets"
          },
          "stack_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Size"
          },
          "max_stack_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Stack Size"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Item"
      },
      "ItemProperty": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ItemProperty"
      },
...
      "RefreshResponse": {
        "properties": {
          "profile": {
            "type": "boolean",
            "title": "Profile"
          },
          "leagues": {
            "type": "boolean",
            "title": "Leagues"
          },
          "characters": {
            "type": "boolean",
            "title": "Characters"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "profile",
          "leagues",
          "characters"
        ],
        "title": "RefreshResponse"
      },
      "RefreshStashesRequest": {
        "properties": {
          "league": {
            "type": "string",
            "title": "League"
          }
        },
        "type": "object",
        "required": [
          "league"
        ],
        "title": "RefreshStashesRequest"
      },
      "Socket": {
        "properties": {
          "group": {
            "type": "integer",
            "title": "Group",
            "default": 0
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": ""
          }
        },
        "type": "object",
        "title": "Socket"
      },
      "StashColour": {
        "properties": {
          "r": {
            "type": "integer",
            "title": "R",
            "default": 0
          },
          "g": {
            "type": "integer",
            "title": "G",
            "default": 0
          },
          "b": {
            "type": "integer",
            "title": "B",
            "default": 0
          }
        },
        "type": "object",
        "title": "StashColour"
      },
      "StashListResponse": {
        "properties": {
          "league": {
            "type": "string",
            "title": "League"
          },
          "tabs": {
            "items": {
              "$ref": "#/components/schemas/StashTabSummary"
            },
            "type": "array",
            "title": "Tabs"
          }
        },
        "type": "object",
        "required": [
          "league",
          "tabs"
        ],
        "title": "StashListResponse"
      },
      "StashTab": {
        "properties": {
          "tab": {
            "$ref": "#/components/schemas/StashTabSummary"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "tab"
        ],
        "title": "StashTab"
      },
      "StashTabSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "NormalStash"
          },
          "index": {
            "type": "integer",
            "title": "Index",
            "default": 0
          },
          "colour": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StashColour"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StashTabSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}